+++ Soundfingerprinting/ISpectrumService.cs
@@ -13,6 +13,6 @@
 		/// <param name="fingerprintImageLength">Length of 1 fingerprint image (w parameter equal to 128, which alogside overlap leads to 128*64 = 8192 = 1.48s)</param>
 		/// <param name="overlap">Overlap between consecutive spectral images, taken previously (64 ~ 11.6ms)</param>
 		/// <returns>List of logarithmic images</returns>
-		List<float[][]> CutLogarithmizedSpectrum(float[][] logarithmizedSpectrum, IStride strideBetweenConsecutiveImages, int fingerprintImageLength, int overlap);
+		List<double[][]> CutLogarithmizedSpectrum(double[][] logarithmizedSpectrum, IStride strideBetweenConsecutiveImages, int fingerprintImageLength, int overlap);
 	}
 }
